feat: populate changed files when starting monokle
