+++ src/redux/git/git.slice.ts
@@ -31,6 +31,7 @@ export const gitSlice = createSlice({
   extraReducers: builder => {
     builder.addCase(setRootFolder.fulfilled, (state, action) => {
       state.repo = action.payload.gitRepo;
+      state.changedFiles = action.payload.gitChangedFiles;
     });
   },
 });
